specials_extractor_service/requirements.txt
requests
Flask
flask-restful
flask-cors
google-cloud-pubsub==0.35.0
jsonpickle